refactor colour util
